+++ .dumirc.ts
@@ -11,7 +11,8 @@ const isWin = process.platform === 'win32';
 export const packages = ['react', 'plugins', 'renderer'];
 
 const nav: INavItem[] = [
-  { link: '/components', title: 'Components' },
+  { link: '/components/react/editor', title: 'Components' },
+  { link: '/components/renderer', title: 'Renderer' },
   { link: '/components/plugins/common', title: 'Plugins' },
   { link: 'https://ui.lobehub.com', mode: 'override', title: 'UI' },
   { link: 'https://icon.lobehub.com', mode: 'override', title: 'Icons' },